# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Career Quest API",
    "description": "Backend foundation. Recommendation and HR summary responses are explicit stubs. Completion and import return 501 without writes. All API routes require X-Role and X-Employee-Id; these headers are trusted development identities.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/employees": {
      "get": {
        "tags": [
          "hr"
        ],
        "summary": "List Employees",
        "operationId": "list_employees_api_employees_get",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "employee",
                    "hr"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          },
          {
            "name": "x-employee-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Employee-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing identity headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee or event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import": {
      "post": {
        "tags": [
          "hr"
        ],
        "summary": "Import Data",
        "description": "Validate the dataset-shaped payload without persisting it.",
        "operationId": "import_data_api_import_post",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "employee",
                    "hr"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          },
          {
            "name": "x-employee-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Employee-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing identity headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee or event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hr/summary": {
      "get": {
        "tags": [
          "hr"
        ],
        "summary": "Get Summary",
        "description": "Contract stub: null metrics have not been calculated.",
        "operationId": "get_summary_api_hr_summary_get",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "employee",
                    "hr"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          },
          {
            "name": "x-employee-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Employee-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing identity headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee or event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/employees/{id}": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "Get Employee",
        "operationId": "get_employee_api_employees__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "employee",
                    "hr"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          },
          {
            "name": "x-employee-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Employee-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "401": {
            "description": "Missing identity headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee or event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/employees/{id}/recommendations": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "Get Recommendations",
        "description": "Contract stub: no ranking, gap calculation or LLM calls.",
        "operationId": "get_recommendations_api_employees__id__recommendations_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "employee",
                    "hr"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          },
          {
            "name": "x-employee-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Employee-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing identity headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee or event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/employees/{id}/complete": {
      "post": {
        "tags": [
          "employees"
        ],
        "summary": "Complete Event",
        "description": "Contract stub: no completion, history or skill changes are persisted.",
        "operationId": "complete_event_api_employees__id__complete_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "employee",
                    "hr"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          },
          {
            "name": "x-employee-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Employee-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing identity headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee or event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CareerGoal": {
        "properties": {
          "target_role": {
            "type": "string",
            "title": "Target Role"
          },
          "target_grade": {
            "type": "string",
            "enum": [
              "Junior",
              "Middle",
              "Senior",
              "Lead"
            ],
            "title": "Target Grade"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_role",
          "target_grade"
        ],
        "title": "CareerGoal"
      },
      "CompleteRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Event Id"
          },
          "completion_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Completion Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "completion_id"
        ],
        "title": "CompleteRequest"
      },
      "CompletionResponse": {
        "properties": {
          "employee_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Employee Id"
          },
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Event Id"
          },
          "completion_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Completion Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "already_completed",
              "not_implemented"
            ],
            "title": "Status"
          },
          "skills": {
            "patternProperties": {
              "^\\S+$": {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 0.0
              }
            },
            "propertyNames": {
              "maxLength": 128,
              "minLength": 1
            },
            "type": "object",
            "title": "Skills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employee_id",
          "event_id",
          "completion_id",
          "status",
          "skills"
        ],
        "title": "CompletionResponse"
      },
      "DatasetMeta": {
        "properties": {
          "dataset": {
            "type": "string",
            "title": "Dataset",
            "default": "Career Quest"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0"
          },
          "as_of_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DatasetMeta"
      },
      "Employee": {
        "properties": {
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic",
            "default": false
          },
          "employee_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Employee Id"
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "title": "Role"
          },
          "grade": {
            "type": "string",
            "enum": [
              "Junior",
              "Middle",
              "Senior",
              "Lead"
            ],
            "title": "Grade"
          },
          "tenure_months": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tenure Months"
          },
          "skills": {
            "patternProperties": {
              "^\\S+$": {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 0.0
              }
            },
            "propertyNames": {
              "maxLength": 128,
              "minLength": 1
            },
            "type": "object",
            "title": "Skills"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "manager_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^\\S+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manager Id"
          },
          "hire_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hire Date"
          },
          "work_format": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "office",
                  "hybrid",
                  "remote"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Format"
          },
          "preferred_language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "kk",
                  "ru",
                  "en"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Language"
          },
          "career_goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CareerGoal"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_review_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Review Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employee_id",
          "role",
          "grade",
          "tenure_months",
          "skills"
        ],
        "title": "Employee"
      },
      "EmployeesResponse": {
        "properties": {
          "employees": {
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "type": "array",
            "title": "Employees"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employees",
          "total"
        ],
        "title": "EmployeesResponse"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HRSummaryResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "not_implemented"
            ],
            "title": "Status"
          },
          "total_employees": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Employees"
          },
          "total_events": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Events"
          },
          "completed_activities": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Activities"
          },
          "employees_by_grade": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "propertyNames": {
                  "enum": [
                    "Junior",
                    "Middle",
                    "Senior",
                    "Lead"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employees By Grade"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HRSummaryResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportRequest": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/DatasetMeta"
          },
          "employees": {
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "type": "array",
            "title": "Employees"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ActivityHistory"
            },
            "type": "array",
            "title": "History"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employees",
          "history"
        ],
        "title": "ImportRequest"
      },
      "ImportResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "imported",
              "not_implemented"
            ],
            "title": "Status"
          },
          "employees_received": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Employees Received"
          },
          "history_received": {
            "type": "integer",
            "minimum": 0.0,
            "title": "History Received"
          },
          "employees_imported": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Employees Imported"
          },
          "history_imported": {
            "type": "integer",
            "minimum": 0.0,
            "title": "History Imported"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "employees_received",
          "history_received",
          "employees_imported",
          "history_imported"
        ],
        "title": "ImportResponse"
      },
      "Recommendation": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event",
          "score",
          "reasons"
        ],
        "title": "Recommendation"
      },
      "RecommendationsResponse": {
        "properties": {
          "employee_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Employee Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "not_implemented"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "enum": [
              "mock",
              "fallback",
              "llm"
            ],
            "title": "Source"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/SkillGap"
            },
            "type": "array",
            "title": "Gaps"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employee_id",
          "status",
          "source",
          "recommendations",
          "gaps"
        ],
        "title": "RecommendationsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}